requirements.txt
agno==2.0.4
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dataclasses==0.6
distro==1.9.0
Django==3.1.12
djongo==1.3.7
dnspython==2.7.0
docstring_parser==0.17.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.36.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.1
markdown-it-py==4.0.0
mcp==1.14.0
mdurl==0.1.2
packaging==25.0
pdfminer.six==20240706
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==3.11.4
pyparsing==3.2.3
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.2.4
sse-starlette==3.0.2
starlette==0.47.3
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
